+++ green-plate/src/main/java/com/team404x/greenplate/livecommerce/entity/Livecommerce.java
@@ -9,7 +9,7 @@ import java.util.List;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 
-import com.team404x.greenplate.company.entity.Company;
+import com.team404x.greenplate.company.model.entity.Company;
 import com.team404x.greenplate.livecommerce.keyword.LivecommerceKeyword;
 import com.team404x.greenplate.recipe.model.entity.Recipe;
