+++ green-plate/src/main/java/com/team404x/greenplate/user/address/entity/Address.java
@@ -3,7 +3,7 @@ package com.team404x.greenplate.user.address.entity;
 import org.hibernate.annotations.ColumnDefault;
 
 import com.team404x.greenplate.orders.entity.Orders;
-import com.team404x.greenplate.user.entity.User;
+import com.team404x.greenplate.user.model.entity.User;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
