+++ green-plate/src/main/java/com/team404x/greenplate/user/keyword/entity/UserKeyword.java
@@ -1,7 +1,7 @@
 package com.team404x.greenplate.user.keyword.entity;
 
 import com.team404x.greenplate.keyword.entity.Keyword;
-import com.team404x.greenplate.user.entity.User;
+import com.team404x.greenplate.user.model.entity.User;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
