+++ green-plate/src/main/java/com/team404x/greenplate/user/repository/UserRepository.java
@@ -2,7 +2,7 @@ package com.team404x.greenplate.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.team404x.greenplate.user.entity.User;
+import com.team404x.greenplate.user.model.entity.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 	User findUserByEmail(String email);
